correcoes e melhorias do codigo base; inclusao de classe para gerenciamento de mensagens
git-svn-id: svn+ssh://localhost/var/svn/postcanal@90 59c7e291-545d-0410-9381-d2a3ada94e96

+++ trunk/library/AB/Session.php
@@ -10,12 +10,78 @@
 class AB_Session
 {
     /**
+     * Session name
+     *
+     * @var string
+     */
+    private $session_name = null;
+
+    /**
      * Session table name
      *
      * @var string
      */
-    private static $table_name = 'application_session';
+    private static $table_name = 'ab_session';
 
+    /**
+     * Session constructor
+     *
+     * @param   string  $name
+     * @return  void
+     */
+    public function __construct($name='default')
+    {
+        session_start();
+        $this->session_name = $name;
+    }
+
+    /**
+     * As of PHP 5.0.5 the write  and close  handlers are called after object
+     * destruction and therefore cannot use objects or throw exceptions. The object
+     * destructors can however use sessions.
+     * 
+     * It is possible to call session_write_close() from the destructor to
+     * solve this chicken and egg problem. 
+     *
+     * @see http://br2.php.net/manual/en/function.session-set-save-handler.php
+     */
+    public function __destruct() { session_write_close(); }
+
+    /**
+     * Get overloading (session attribute)
+     *
+     * @param   string  $name
+     * @return  mixed
+     */
+    protected function __get($name)
+    {
+        $value = null;
+
+        if(array_key_exists($this->session_name, $_SESSION))
+        {
+            if(array_key_exists($name, $_SESSION[$this->session_name]))
+            {
+                $value = $_SESSION[$this->session_name][$name];
+            }
+        }
+
+        return $value;
+    }
+
+    /**
+     * Set overloading (session attribute)
+     *
+     * @param   string  $name
+     * @param   mixed   $value
+     * @return  void
+     */
+    protected function __set($name, $value)
+    {
+        if(is_array($_SESSION[$this->session_name]))
+        {
+            $_SESSION[$this->session_name][$name] = $value;
+        }
+    }
 
     /**
      * Open the session
@@ -40,21 +106,21 @@ class AB_Session
     /**
      * Read the session
      *
-     * @param   integer     $id
+     * @param   string  $id
      * @return  string
      */
     public static function read ($id)
     {
         $result = AB_Model::selectRow("SELECT session_data " . 
                                       "FROM " . self::$table_name . " " .
-                                      "WHERE session = ?", array($id));
+                                      "WHERE id = ?", array($id));
         return is_object($result) ? $result->session_data : '';
     }
 
     /**
      * Write the session
      *
-     * @param   integer     $id
+     * @param   string      $id
      * @param   string      $data
      * @return  boolean
      */
@@ -62,15 +128,15 @@ class AB_Session
     {
         $result = AB_Model::selectRow("SELECT COUNT(*) AS total " . 
                                       "FROM " . self::$table_name . " " .
-                                      "WHERE session = ?", array($id));
+                                      "WHERE id = ?", array($id));
 
         return AB_Model::execute
         (
             ($result->total > 0) ? 
                 "UPDATE " . self::$table_name . " SET " .
                 "session_expires = ?, session_data = ? " .
-                "WHERE session = ?" :
-                "INSERT INTO " . self::$table_name . " VALUES (?, ?, ?)",
+                "WHERE id = ?" :
+                "INSERT INTO " . self::$table_name . " VALUES (?, ?, ?, 0)",
             ($result->total > 0) ?
                 array(time(), $data, $id) :
                 array($id, time(), $data)
@@ -80,13 +146,12 @@ class AB_Session
     /**
      * Destroy the session
      *
-     * @param   integer     $id
-     * @return bool
+     * @param   string  $id
+     * @return  boolean
      */
     public static function destroy ($id) 
     {
-        return AB_Model::execute("DELETE FROM " . self::$table_name . " " .
-                                 "WHERE session = ?",
+        return AB_Model::execute("DELETE FROM " . self::$table_name . " WHERE id = ?",
                                  array($id));
     }
 
@@ -94,13 +159,24 @@ class AB_Session
      * Garbage Collector
      *
      * @param   integer     $life 
-     * @return  bool
+     * @return  boolean
      */
-    public static function gc ($life) 
+    public static function gc ($max) 
     {
+        $registry = AB_Registry::singleton();
+        $expiration = $registry->session->expiration;
+
+        if(intval($expiration) <= 0)
+        {
+            $_m = "session expiration value must be greater than zero";
+            $_d = array('method' => __METHOD__);
+            throw new AB_Exception($_m, E_USER_ERROR, $_d);
+        }
+
         return AB_Model::execute("DELETE FROM " . self::$table_name . " " .
-                                 "WHERE session_expires < ?",
-                                 array(time() - $life));
+                                 "WHERE (session_expires < ? AND session_active = 0) " .
+                                 "OR (session_expires < ? AND session_active = 1)",
+                                 array((time() - $max), (time() - $expiration)));
     }
 
     /**
@@ -119,5 +195,42 @@ class AB_Session
                                  array('AB_Session', 'destroy'),
                                  array('AB_Session', 'gc')
         );
+    }
+
+    /**
+     * Switch session activity
+     *
+     * @param   boolean     $active
+     * @return  boolean
+     */
+    public function setActive($active=true)
+    {
+        if($active == true)
+        {
+            $_SESSION[$this->session_name] = array();
+        }
+        else
+        {
+            unset($_SESSION[$this->session_name]);
+        }
+
+        return AB_Model::execute("UPDATE " . self::$table_name . " " .
+                                 "SET active = ? WHERE id = ?",
+                                 array($active, session_id()));
+    }
+
+    /**
+     * Get session activity
+     *
+     * @return  boolean
+     */
+    public function getActive()
+    {
+        $result = AB_Model::selectRow("SELECT COUNT(*) AS total " . 
+                                      "FROM " . self::$table_name . " " .
+                                      "WHERE id = ? AND active = 1", 
+                                      array(session_id()));
+
+        return ($result->total > 0);
     }
 }
